requirements.txt
requests
urlopen
BeautifulSoup
pandas
geocoder
os
mysql.connector
sqlalchemy